backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator<2.0.0
bcrypt==4.1.2
python-dotenv==1.0.0

# News fetching
requests==2.31.0
feedparser==6.0.11
gnews==0.3.7
python-dateutil==2.8.2

# LLM - Alibaba Cloud Qwen
dashscope==1.14.1

# LLM - OpenAI compatible (for NVIDIA GLM API)
openai>=1.0.0

# Scheduler
apscheduler==3.10.4

# Email
resend==0.7.0

# Database
alembic==1.13.1
